adding logic of accounting entries from cash receipts

+++ src/main/java/com/account_catalogue/catalogue/domain/models/ReceiptDetail.java
@@ -1,5 +1,7 @@
 package com.account_catalogue.catalogue.domain.models;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -10,8 +12,7 @@ import lombok.Setter;
 public class ReceiptDetail {
     private Long id;
     private Long originalInvoiceId;
-    private Long amountPaid;
+    private BigDecimal amountPaid;
     private String invoiceCode;
     private Long accountingAccount;
-    
 }
